Move reduce and while printing to AssemblyFormat to be shared with MHLO
